WIP - Refactor account web
