+++ account/web/app/src/scenes/Billing/components/PaymentMethod/PaymentMethod.scss
@@ -17,12 +17,7 @@
         line-height: 22px;
 
         span {
-          margin-right: 10px;
-          font-size: 22pt;
-          display: inline-block;
-          vertical-align: middle;
-          line-height: normal;
-          padding-bottom: 2px;
+          text-transform: capitalize;
         }
       }
