+++ account/web/app/src/scenes/Billing/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Billing';
