+++ account/web/app/src/scenes/Profile/Profile.scss
@@ -1,0 +1,48 @@
+.Profile {
+  form {
+    display: flex;
+    text-align: left;
+    flex-direction: column;
+    flex-grow: 1;
+    margin-top: 20px;
+
+    label {
+      margin-bottom: 8px;
+      font-size: 10pt;
+    }
+
+    input{
+      border-radius: 4px;
+      outline: none;
+    }
+
+    .StripeElement {
+      border-radius: 4px;
+      margin-bottom: 25px;
+      padding: 9px 10px;
+      border: solid 1px #00000020;
+    }
+
+    input[type=text], input[type=email] {
+      height: 30px;
+      margin-bottom: 25px;
+      padding: 3px 10px;
+      border: solid 1px #00000020;
+    }
+
+    input[type=submit] {
+      margin-top: 10px;
+      height: 50px;
+      background-color: #66AC5C;
+      font-weight: bold;
+      color: white;
+      border: none;
+      font-size: 10pt;
+      cursor: pointer;
+
+      &:disabled {
+        background-color: lightgray;
+      }
+    }
+  }
+}
