+++ account/web/app/src/scenes/Profile/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Profile';
